This document will lead you in the installation of Google Chrome for Windows, to be used as client for the Chronos Web Modeler (CWM). CWM is a collaborative tool for the creation of UML models. See https://sourceforge.net/projects/olympos/ for more information.
In some company very old version of the browser are installed. CWM requires a browser with Just In time compiler to works smoothly. 
The solution  is to install a modern browser like Goggle Chrome.
Google Chrome can transform it self in an "application" on the desktop that will execute CWM on